Draw a line in the sand after a long time
Draw a line in the sand after a long time

This commit is just a commit of the flextime.app and all other stuff
that is in here without really looking or knowing what's changed or
added.

I have been running the flextime.app code for several months and it is
useful to me on a daily basis.  I have not been looking at the code
since July and it is a bit of a mess.  I left it when I realized that
I just wanted to wait for async/await instead of coding like an
animal..

So instead of trying to make sens of it, I am just checking the
current situation in.