Clear measurements between file writes
This commit will fix a problem where we did not clear measurements
between writing data to files.  This meant that measurements
accumulated and caused big performance problems.

This commit also updates the way we treat files.  We now overrwrite
the existing file up to a time limit, before we write to a new file.
This means that every measurement immediately gets written to file,
and is no longer kept in memory only.

This also means that it is no longer as important to capture SIGINT,
but it is still there.