Create measurements at one minute marks
This commit will create measurements as close to the minute as
possble.  The reason is that when the user runs a client to inspect
working hours, we want the end time to match the clock.  Since we
typically resolve down to a minute, we want each measurement _and_
flush to disk to happen _just_ as a new minute on the clock starts.